refactor: fix name conflicts
